@@ -12,25 +12,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.7",
-    "@angular/common": "^4.4.7",
-    "@angular/compiler": "^4.4.7",
-    "@angular/core": "^4.4.7",
-    "@angular/forms": "^4.4.7",
-    "@angular/http": "^4.4.7",
-    "@angular/platform-browser": "^4.4.7",
-    "@angular/platform-browser-dynamic": "^4.4.7",
-    "@angular/router": "^4.4.7",
+    "@angular/animations": "^5.0.0",
+    "@angular/common": "^5.0.0",
+    "@angular/compiler": "^5.0.0",
+    "@angular/core": "^5.0.0",
+    "@angular/forms": "^5.0.0",
+    "@angular/http": "^5.0.0",
+    "@angular/platform-browser": "^5.0.0",
+    "@angular/platform-browser-dynamic": "^5.0.0",
+    "@angular/platform-server": "^5.0.0",
+    "@angular/router": "^5.0.0",
     "core-js": "^2.5.7",
     "font-awesome": "^4.7.0",
     "ng-lazyload-image": "^3.4.2",
     "primeng": "^4.3.0",
-    "rxjs": "^5.5.11",
+    "rxjs": "^5.5.2",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular/cli": "1.2.6",
-    "@angular/compiler-cli": "^4.4.7",
+    "@angular/compiler-cli": "^5.0.0",
     "@angular/language-service": "^4.4.7",
     "@types/jasmine": "^2.5.54",
     "@types/jasminewd2": "^2.0.3",
@@ -47,6 +48,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "^2.4.2"
   }
 }
