@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
 import { CarService } from './car.service';
 import {SelectItem,MenuItem,LazyLoadEvent,FilterMetadata} from './api';
@@ -8,7 +8,7 @@ import {SelectItem,MenuItem,LazyLoadEvent,FilterMetadata} from './api';
 @NgModule({
   imports: [
     CommonModule,
-    HttpModule
+    HttpClientModule
   ],
   
   providers: [CarService]
