@@ -1,5 +1,5 @@
 import { NgModule,ModuleWithProviders } from '@angular/core';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
 import { CarService } from './car.service';
 
@@ -7,7 +7,7 @@ import { CarService } from './car.service';
 @NgModule({
   imports: [
     CommonModule,
-    HttpModule
+    HttpClientModule
   ]
 })
 export class ServiceModule { 
