@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Http, Response, Headers, URLSearchParams, RequestOptions  } from '@angular/http';
+import { HttpClient, HttpResponse, HttpHeaders,  HttpParams  } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { Car,Trabajador} from '../interface';
 //import {Trabajador} from './../interface/trabajador';
@@ -9,26 +9,41 @@ import 'rxjs/add/operator/toPromise';
 @Injectable()
 export class CarService {
     
-    constructor(private http: Http) {}
+    constructor(private http: HttpClient) {
+
+    }
+
+    
+      
+
 
     postPdt(post): Observable<any> {
-        let myHeaders = new Headers();
-        myHeaders.append( 'Content-Type', 'application/json' );
-        let options = new RequestOptions({ headers: myHeaders });
-        return this.http.post('/api/cas/pdt_cas',JSON.stringify(post) , options )
-        .map((response: Response) => {     
-            return <any>response.json();
+        const headers = new HttpHeaders();
+        headers.append('Content-Type', 'application/json');
+
+       //let params = new HttpParams();
+       //params = params.append('dni', '40148011');
+       
+       
+
+        //let options = new RequestOptions({ headers: myHeaders });
+        return this.http.post('/api/cas/pdt_cas',post, {headers} )
+        .map((response: HttpResponse<any>) => {     
+            return response;
         })
         .catch(this.onError);
     }
    
     postTodo(post): Observable<any> {
-        let myHeaders = new Headers();
-        myHeaders.append( 'Content-Type', 'application/json' );
-        let options = new RequestOptions({ headers: myHeaders });
-        return this.http.post('api/todo',JSON.stringify(post) , options )
-        .map((response: Response) => {     
-            return <any>response.json();
+        const headers = new HttpHeaders();
+        headers.append('Content-Type', 'application/json');
+
+       let params = new HttpParams();
+       params = params.append('param-1', 'value-1');
+
+        return this.http.post('api/todo',JSON.stringify(post) , {headers} )
+        .map((response: HttpResponse<any>) => {     
+            return <any>response;
         })
         .catch(this.onError);
     }
@@ -43,8 +58,8 @@ export class CarService {
         console.log('recibiendo '+dni);
          return this.http
              .get('http://programacion.minsa/api/trabajador/'+dni)
-             .map((response: Response) => {     
-                 return <any[]>response.json();
+             .map((response: HttpResponse<any[]>) => {     
+                 return  <any>response;
              })
              .catch(this.handleError);
      }
@@ -52,8 +67,8 @@ export class CarService {
      deleteTodo(id): Observable<any> {        
          return this.http
              .delete('http://programacion.minsa/api/todo/'+id)
-             .map((response: Response) => {     
-                 return <any>response.json();
+             .map((response: HttpResponse<any>) => {     
+                 return <any>response;
              })
              .catch(this.handleError);
      }
@@ -61,41 +76,41 @@ export class CarService {
     getTodo(): Observable<any[]> {        
          return this.http
              .get('/api/todo')
-             .map((response: Response) => {     
-                 return <any[]>response.json();
-             })
+             .map((response: HttpResponse<any[]>) => {     
+                return  <any>response;
+            })
              .catch(this.handleError);
      }
  
-     private handleError(error: Response) {
+     private handleError(error: HttpResponse<any>) {
          return Observable.throw(error.statusText);
      }
 
     getImgTrabajador(dni) {
         return this.http.get('http://programacion.minsa/api/trabajador/'+dni)
                     .toPromise()
-                    .then(res => <Car[]> res.json())
+                    .then(res => <Car[]> res)
                     .then(data => { return data; });
     }
 
     getCarsSmall() {
         return this.http.get('http://programacion.minsa/api/cas/analisis')
                     .toPromise()
-                    .then(res => <Car[]> res.json())
+                    .then(res => <Car[]> res)
                     .then(data => { return data; });
     }
 
     getCarsMedium() {
         return this.http.get('assets/data/cars-small.json')
                     .toPromise()
-                    .then(res => <any[]> res.json().data)
+                    .then(res => <any[]> res)
                     .then(data => { return data; });
     }
 
     getCarsLarge(page,campo,orden) {
         return this.http.get('http://programacion.minsa/api/alta_vigente/'+campo+'/'+orden+'?page='+page)
                     .toPromise()
-                    .then(res => <Trabajador[]> res.json())
+                    .then(res => <Trabajador[]> res)
                     .then(data => { return data; });
     }
 }
